+++ packages/authentication/jest.config.ts
@@ -1,6 +1,1 @@
-const jestConfigs = require('@baseapp-frontend/test/jest.config.ts')
-
-module.exports = {
-  ...jestConfigs,
-  testMatch: [...(jestConfigs.testMatch ?? []), '<rootDir>/**/*.(spec|test).(ts|tsx)'],
-}
+module.exports = require('@baseapp-frontend/test/jest.config.ts')