+++ packages/authentication/modules/profile/useCurrentProfile/__tests__/useCurrentProfile.test.tsx
@@ -11,7 +11,7 @@ jest.mock('@baseapp-frontend/utils', () => ({
   setCookie: jest.fn(),
 }))
 
-describe('CurrentProfileProvider', () => {
+describe('useCurrentProfile', () => {
   afterEach(() => {
     jest.clearAllMocks()
   })